+++ kafka/versions.py
@@ -6,7 +6,7 @@ versions = [
         "scala": "2.13",
         "kafka/kcat": "1.7.0",
         "opa_authorizer": "1.5.1",
-        "jmx_exporter": "1.2.0",
+        "jmx_exporter": "1.3.0",
     },
     {
         "product": "3.9.0",
@@ -15,7 +15,7 @@ versions = [
         "scala": "2.13",
         "kafka/kcat": "1.7.0",
         "opa_authorizer": "1.5.1",
-        "jmx_exporter": "1.2.0",
+        "jmx_exporter": "1.3.0",
     },
     {
         "product": "3.9.1",
@@ -24,7 +24,7 @@ versions = [
         "scala": "2.13",
         "kafka/kcat": "1.7.0",
         "opa_authorizer": "1.5.1",
-        "jmx_exporter": "1.2.0",
+        "jmx_exporter": "1.3.0",
     },
     {
         "product": "4.0.0",
@@ -33,6 +33,6 @@ versions = [
         "scala": "2.13",
         "kafka/kcat": "1.7.0",
         "opa_authorizer": "1.5.1",
-        "jmx_exporter": "1.2.0",
+        "jmx_exporter": "1.3.0",
     },
 ]